refactor: split projects into core project and plugin projects
Add solution file for better management
