+++ OsuRTDataProvider.Plugin/Language/en-US/OsuRTDataProvider.Plugin.PluginLanguage.lang
