+++ OsuRTDataProvider.Plugin/Language/ja-JP/OsuRTDataProvider.Plugin.PluginLanguage.lang
