+++ OsuRTDataProvider.Plugin/Language/zh-CN/OsuRTDataProvider.Plugin.PluginLanguage.lang
