+++ OsuRTDataProvider.Plugin/SettingIni.cs
@@ -1,8 +1,6 @@
-﻿using Sync.Tools.ConfigurationAttribute;
 using Sync.Tools;
+using Sync.Tools.ConfigurationAttribute;
 using System;
-using System.Linq;
-using System.Text.RegularExpressions;
 
 namespace OsuRTDataProvider
 {
@@ -82,25 +80,5 @@ namespace OsuRTDataProvider
         public void onConfigurationSave()
         {
         }
-    }
-
-    internal static class Setting
-    {
-        public static bool DebugMode = false;
-        public static int ListenInterval = 100;//ms
-        public static bool EnableTourneyMode = false;
-        public static int TeamSize = 1;
-        public static string ForceOsuSongsDirectory = "";
-        public static string GameMode = "Auto";
-        public static bool DisableProcessNotFoundInformation = false;
-        public static bool EnableModsChangedAtListening = false;
-
-        #region NoSave
-        public static string SongsPath = string.Empty;
-        public static string OsuVersion = string.Empty;
-        public static string Username = string.Empty;
-        #endregion
-
-        public static double CurrentOsuVersionValue => Utils.ConvertVersionStringToValue(OsuVersion);
     }
 }
