+++ OsuRTDataProvider/DefaultLanguage.cs
@@ -1,0 +1,33 @@
+namespace OsuRTDataProvider
+{
+    public static class DefaultLanguage
+    {
+        public static string LANG_OSU_NOT_FOUND = "[ID:{0}]Not found osu!.exe process";
+        public static string LANG_OSU_FOUND = "[ID:{0}]Found osu!.exe process";
+        public static string LANG_TOURNEY_HINT = "Tourney Mode: {0}";
+
+        public static string CHECK_GOTO_RELEASE_PAGE_HINT = "Enter \"ortdp releases\" to open the releases page in your browser.";
+        public static string LANG_CHECK_ORTDP_UPDATE = "Found a new version of OsuRTDataProvider({0})";
+        public static string LANG_INIT_STATUS_FINDER_FAILED = "[ID:{0}]Init StatusFinder Failed! Retry after {1} seconds";
+        public static string LANG_INIT_STATUS_FINDER_SUCCESS = "[ID:{0}]Init StatusFinder Success!";
+        public static string LANG_INIT_PLAY_FINDER_FAILED = "[ID:{0}]Init PlayFinder Failed! Retry after {1} seconds";
+        public static string LANG_INIT_PLAY_FINDER_SUCCESS = "[ID:{0}]Init PlayFinder Success!";
+        public static string LANG_INIT_BEATMAP_FINDER_FAILED = "[ID:{0}]Init BeatmapFinder Failed! Retry after {1} seconds";
+        public static string LANG_INIT_BEATMAP_FINDER_SUCCESS = "[ID:{0}]Init BeatmapFinder Success!";
+        public static string LANG_INIT_MODE_FINDER_FAILED = "[ID:{0}]Init ModeFinder Failed! Retry after {1} seconds";
+        public static string LANG_INIT_MODE_FINDER_SUCCESS = "[ID:{0}]Init ModeFinder Success!";
+        public static string LANG_INIT_HIT_EVENT_SUCCESS = "[ID:{0}]Init HitEventFinder Success!";
+        public static string LANG_INIT_HIT_EVENT_FAIL = "[ID:{0}]Init HitEventFinder Failed! Retry after {1} seconds";
+
+        public static string LANG_BEATMAP_NOT_FOUND = "Beatmap not found";
+
+        public static string ListenInterval = "Listen interval(ms)";
+        public static string EnableTourneyMode = "Tourney mode";
+        public static string TeamSize = "Team size";
+        public static string DebugMode = "Debug mode";
+        public static string ForceOsuSongsDirectory = "Force OSU! songs directory";
+        public static string GameMode = "Game Mode";
+        public static string DisableProcessNotFoundInformation = "Disable OSU! process not found information";
+        public static string EnableModsChangedAtListening = "Enable Mods Changed At Listening(Experimental)";
+    }
+}
