+++ OsuRTDataProvider/Helper/PathHelper.cs
