+++ OsuRTDataProvider/Memory/BeatmapOffsetInfo.cs
