+++ OsuRTDataProvider/Memory/LogHelper.cs
@@ -1,4 +1,3 @@
-﻿using Sync.Tools;
 using System;
 using System.Collections.Generic;
 using System.Linq;
