+++ OsuRTDataProvider/Memory/OsuFinderBase.cs
@@ -1,5 +1,4 @@
-﻿using Sync.Tools;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Runtime.InteropServices;
