+++ OsuRTDataProvider/Memory/OsuInternalStatus.cs
