+++ OsuRTDataProvider/Memory/OsuModeFinder.cs
