+++ OsuRTDataProvider/Memory/OsuStatusFinder.cs
