+++ OsuRTDataProvider/Memory/SigScan.cs
