+++ OsuRTDataProvider/Mods/ModsInfo.cs
