+++ OsuRTDataProvider/VersionInfo.cs
@@ -1,0 +1,10 @@
+﻿using System.Reflection;
+
+namespace OsuRTDataProvider
+{
+    internal static class VersionInfo
+    {
+        public static string GetVersion() => Assembly.GetExecutingAssembly()
+            .GetCustomAttribute<AssemblyInformationalVersionAttribute>().InformationalVersion;
+    }
+}
